dataset-pipeline/requirements.txt
numpy==1.26.4
pyarrow==16.0.0
psycopg2-binary==2.9.9
clickhouse-driver==0.2.7
minio==7.2.7
